ConcreteInterpreter

Description

Concrete interpreter of the While language.

Documentation

data Val

Values of the While language can be booleans or numbers.

Constructors

BoolVal Bool

NumVal Int

type Addr = Label

run :: [LStatement] -> Error String (HashMap Addr Val)

The concrete interpreter of the while language instantiates run with the concrete components for failure (FailureT), store (StoreT), environments (EnvT), random numbers (RandomT), and values (ConcreteT).

newtype ConcreteT c x y

The ConcreteT transformer defines the value operations for the While language.

Constructors

ConcreteT

Fields

runConcreteT :: c x y