Exceptions.ConcreteInterpreter

Description

Concrete interpreter of the While language.

Documentation

type Exception = (Text, Val)

Values of the While language can be booleans or numbers.

run :: [LStatement] -> Error String (Error Exception (HashMap Addr Val))

The concrete interpreter of the while language instantiates run with the concrete components for failure (FailureT), store (StoreT), environments (EnvT), random numbers (RandomT), and values (ConcreteT).

Orphan instances

ArrowChoice c => IsException Exception Val (ConcreteT c)

Instance details

Associated Types

type JoinExc (ConcreteT c) x y :: Constraint

Methods

namedException :: ConcreteT c (Text, Val) Exception

matchException :: JoinExc (ConcreteT c) ((Val, x), x) y => ConcreteT c (Val, x) y -> ConcreteT c x y -> ConcreteT c (Text, Exception, x) y