Exceptions.GenericInterpreter

Description

Generic interpreter for the While-Language.

Documentation

type Prog = [Statement]

eval :: (Show addr, ArrowChoice c, ArrowRand v c, ArrowEnv Text addr env c, ArrowStore addr v c, ArrowFail err c, IsString err, ArrowExcept exc c, IsVal v c, IsException exc v c, Join c ((addr, Text), Text) v, Join c ((v, addr), addr) v) => c Expr v

Generic interpreter for expressions of the While-language. It is an arrow computation c Expr v that consumes expressions and produces values. The interpreter is parameterized by the type of values v, addresses addr, environment env and arrow type c. It uses the IsVal interface to combine values and uses the ArrowEnv and ArrowStore interface to access the environment.

run :: (Show addr, ArrowChoice c, ArrowFix [Statement] () c, ArrowEnv Text addr env c, ArrowStore addr v c, ArrowAlloc (Text, v, Label) addr c, ArrowFail err c, ArrowExcept exc c, ArrowRand v c, IsString err, IsVal v c, IsException exc v c, Join c ((addr, Text), Text) v, Join c ((addr, (Text, v, Label)), (Text, v, Label)) addr, Join c ((v, addr), addr) v, Join c ((), ((Statement, (Text, Text, Statement, Label)), exc)) (), Join c (((Statement, Statement), ()), ((Statement, Statement), exc)) (), JoinVal c ([Statement], [Statement]) (), JoinExc c ((v, (Text, Statement, Label)), (Text, Statement, Label)) ()) => c [Statement] ()

Generic interpreter for statements of the While-language. It is an arrow computation c [Statement] () that consumes statements and does not produces a result. The interpreter is parameterized by the type of values v, addresses addr, environment env and arrow type c.

class IsException exc v c | c -> v where

Associated Types

type JoinExc (c :: * -> * -> *) x y :: Constraint

Methods

namedException :: c (Text, v) exc

matchException :: JoinExc c ((v, x), x) y => c (v, x) y -> c x y -> c (Text, exc, x) y