GenericInterpreter

Description

Generic interpreter for the While-Language.

Documentation

type Prog = [Statement]

eval :: (Show addr, ArrowChoice c, ArrowRand v c, ArrowEnv Text addr env c, ArrowStore addr v c, ArrowFail e c, IsString e, IsVal v c, Join c ((addr, Text), Text) v, Join c ((v, addr), addr) v) => c Expr v

Generic interpreter for expressions of the While-language. It is an arrow computation c Expr v that consumes expressions and produces values. The interpreter is parameterized by the type of values v, addresses addr, environment env and arrow type c. It uses the IsVal interface to combine values and uses the ArrowEnv and ArrowStore interface to access the environment.

run :: (Show addr, ArrowChoice c, ArrowFix [Statement] () c, ArrowEnv Text addr env c, ArrowStore addr v c, ArrowAlloc (Text, v, Label) addr c, ArrowFail e c, ArrowRand v c, IsString e, IsVal v c, Join c ((addr, Text), Text) v, Join c ((addr, (Text, v, Label)), (Text, v, Label)) addr, Join c ((v, addr), addr) v, JoinVal c ([Statement], [Statement]) ()) => c [Statement] ()

Generic interpreter for statements of the While-language. It is an arrow computation c [Statement] () that consumes statements and does not produces a result. The interpreter is parameterized by the type of values v, addresses addr, environment env and arrow type c.

class Arrow c => IsVal v c | c -> v where

Interface for value operations.

Associated Types

type JoinVal (c :: * -> * -> *) x y :: Constraint

This constraint allows the abstract interpreter to join the results of an if statement.

Methods

boolLit :: c (Bool, Label) v

and :: c (v, v, Label) v

or :: c (v, v, Label) v

not :: c (v, Label) v

numLit :: c (Int, Label) v

add :: c (v, v, Label) v

sub :: c (v, v, Label) v

mul :: c (v, v, Label) v

div :: c (v, v, Label) v

eq :: c (v, v, Label) v

lt :: c (v, v, Label) v

if_ :: JoinVal c (x, y) z => c x z -> c y z -> c (v, (x, y)) z