ReachingDefinitionsAnalysis

Description

Reaching Definition Analysis for the While language.

Documentation

run :: (?bound :: IV) => Int -> [LStatement] -> [(Label, Map Text (Pow Label))]

Calculates the entry sets of which definitions may be reached for each statment. The analysis instantiates the generic interpreter run with analysis components for fixpoint computation (FixT), termination (TerminatingT), failure (ErrorT), store (StoreT), environments (EnvT), and values (IntervalT)

combineMaps :: (Identifiable k, Identifiable a) => Map k a -> Map a v -> Map k v

dropValues :: Map a (v, l) -> Map a l

joinOnKey :: (Identifiable k', Complete v') => (k -> v -> Maybe (k', v')) -> HashMap k v -> HashMap k' v'

Orphan instances

HasLabel (x, [Statement]) Label

Instance details

Methods

label :: (x, [Statement]) -> Label